add upgrade
add infinity upgrade
add stack upgrade starter tier
add try remove stack upgrade fail warning
change stack upgrade multiplier to double
fix jukebox upgrade no close gui bug
fix take stack bug

+++ src/main/java/ruiseki/okbackpack/api/wrapper/IStackSizeUpgrade.java
@@ -2,6 +2,6 @@ package ruiseki.okbackpack.api.wrapper;
 
 public interface IStackSizeUpgrade extends ISlotModifiable {
 
-    int getMultiplier();
+    double getMultiplier();
 
 }
